LayoutTests/http/tests/security/cross-origin-appcache-allowed-expected.txt

CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
This iframe should have a cache:


This iframe should have a cache:



--------
Frame: '<!--frame1-->'
--------
Cache found

--------
Frame: '<!--frame2-->'
--------
Cache found
